% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_sfnetworks.R
\name{tmapGetShapeMeta1.sfnetwork}
\alias{tmapGetShapeMeta1.sfnetwork}
\title{Implementation of tmapGetShapeMeta1 for sfnetwork}
\usage{
\method{tmapGetShapeMeta1}{sfnetwork}(shp, o)
}
\description{
Implementation of tmapGetShapeMeta1 for sfnetwork
}
\keyword{internal}
