% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_sfnetworks.R
\name{tmapGetShapeMeta2.sfnetwork}
\alias{tmapGetShapeMeta2.sfnetwork}
\title{Implementation of tmap::tmapGetShapeMeta2 for sfnetwork}
\usage{
\method{tmapGetShapeMeta2}{sfnetwork}(shp, smeta, o)
}
\description{
Implementation of tmap::tmapGetShapeMeta2 for sfnetwork
}
\keyword{internal}
