% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_sfnetworks.R
\name{tmapShape.sfnetwork}
\alias{tmapShape.sfnetwork}
\title{Implementation of tmap::tmapShape for sfnetwork}
\usage{
\method{tmapShape}{sfnetwork}(shp, is.main, crs, bbox, unit, filter, shp_name, smeta, o, tmf)
}
\description{
Implementation of tmap::tmapShape for sfnetwork
}
\keyword{internal}
