% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_sfnetworks.R
\name{tmapSubsetShp.sfnetwork}
\alias{tmapSubsetShp.sfnetwork}
\title{Implementation of tmapSubsetShp for sfnetwork}
\usage{
\method{tmapSubsetShp}{sfnetwork}(shp, vars)
}
\description{
Implementation of tmapSubsetShp for sfnetwork
}
\keyword{internal}
